refactor: move all shared components to common folder
