+++ src/components/common/WebhookViewer.tsx
@@ -1,5 +1,5 @@
-import { Badge } from "./ui/badge";
-import { Button } from "./ui/button";
+import { Badge } from "../ui/badge";
+import { Button } from "../ui/button";
 
 import ConnectButton from "./ConnectButton";
 import MessageCard from "./MessageCard";
